fixing remaining (mostly compilation) issues after initial rebase of cep-15-accord on cep-21-tcm

+++ src/java/org/apache/cassandra/db/rows/AbstractCell.java
@@ -120,7 +120,7 @@ public abstract class AbstractCell<V> extends Cell<V>
     @Override
     public ColumnData updateAllTimestampAndLocalDeletionTime(long newTimestamp, int newLocalDeletionTime)
     {
-        int localDeletionTime = localDeletionTime() != NO_DELETION_TIME ? newLocalDeletionTime : NO_DELETION_TIME;
+        long localDeletionTime = localDeletionTime() != NO_DELETION_TIME ? newLocalDeletionTime : NO_DELETION_TIME;
         return new BufferCell(column, isTombstone() ? newTimestamp - 1 : newTimestamp, ttl(), localDeletionTime, buffer(), path());
     }
